Rename PlotlyDarkTheme to PlotlyCloroTemplates
